feat(detector-aws): add EC2 detector implementing DetectorSync interface
