+++ detectors/node/opentelemetry-resource-detector-aws/src/detectors/index.ts
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-export * from './AwsEc2Detector';
+export { awsEc2Detector } from './AwsEc2Detector';
+export { awsEc2DetectorSync } from './AwsEc2DetectorSync';
 export * from './AwsBeanstalkDetector';
 export * from './AwsEcsDetector';
 export * from './AwsEksDetector';
